+++ mobx/lib/src/api/extensions/primitive_types_extensions.dart
@@ -1,20 +1,20 @@
 part of '../extensions.dart';
 
 extension IntExtension on int {
-  /// turns an int into ObservableInt
-  ObsInt obs({ReactiveContext? context, String? name}) {
+  /// turns an int into Observable<int>
+  Observable<int> obs({ReactiveContext? context, String? name}) {
     return Observable(this, context: context, name: name);
   }
 }
 
 extension BoolExtension on bool {
-  /// turns a bool into ObservableBool
-  ObsBool obs({ReactiveContext? context, String? name}) {
+  /// turns a bool into Observable<bool>
+  Observable<bool> obs({ReactiveContext? context, String? name}) {
     return Observable(this, context: context, name: name);
   }
 }
 
-extension ObservableBoolExtension on ObsBool {
+extension ObservableBoolExtension on Observable<bool> {
   /// lets you toggle the internal value of ObservableBool
   void toggle() {
     runInAction(() => value = !value);
@@ -22,15 +22,15 @@ extension ObservableBoolExtension on ObsBool {
 }
 
 extension DoubleExtension on double {
-  /// turns a double into ObservableDouble
-  ObsDouble obs({ReactiveContext? context, String? name}) {
+  /// turns a double into Observable<double>
+  Observable<double> obs({ReactiveContext? context, String? name}) {
     return Observable(this, context: context, name: name);
   }
 }
 
 extension StringExtension on String {
-  /// turns a String into ObservableString
-  ObsString obs({ReactiveContext? context, String? name}) {
+  /// turns a String into Observable<String>
+  Observable<String> obs({ReactiveContext? context, String? name}) {
     return Observable(this, context: context, name: name);
   }
 }