+++ mobx/lib/src/core/typedefs.dart
@@ -1,7 +1,0 @@
-part of '../core.dart';
-
-/// aliases for primitive types
-typedef ObsInt = Observable<int>;
-typedef ObsDouble = Observable<double>;
-typedef ObsBool = Observable<bool>;
-typedef ObsString = Observable<String>;