+++ mobx/test/extensions/primitive_types_extensions_test.dart
@@ -9,47 +9,28 @@ void main() {
   group('PrimitiveTypesExtensions', () {
     test('Transform Int into ObsInt', () {
       final count = 0;
-      expect(count.obs(), isA<ObsInt>());
+      expect(count.obs(), isA<Observable<int>>());
     });
 
     test('Transform Double into ObsDouble', () {
       final count = 0.0;
-      expect(count.obs(), isA<ObsDouble>());
+      expect(count.obs(), isA<Observable<double>>());
     });
 
     test('Transform Bool into ObsBool', () {
       final flag = false;
-      expect(flag.obs(), isA<ObsBool>());
+      expect(flag.obs(), isA<Observable<bool>>());
     });
 
     test('Transform String into ObsString', () {
       final str = '';
-      expect(str.obs(), isA<ObsString>());
+      expect(str.obs(), isA<Observable<String>>());
     });
 
     test('Toggles ObservableBool', () {
       final flag = false.obs();
       flag.toggle();
       expect(flag.value, equals(true));
-    });
-
-  });
-
-  group('Typedefs', () {
-    test('ObsInt is Observable<int>', () {      
-      expect(ObsInt, isA<Observable<int>>());
-    });
-
-    test('ObsBool is Observable<bool>', () {      
-      expect(ObsBool, isA<Observable<bool>>());
-    });
-
-    test('ObsDouble is Observable<double>', () {      
-      expect(ObsDouble, isA<Observable<double>>());
-    });
-
-    test('ObsString is Observable<String>', () {      
-      expect(ObsString, isA<Observable<String>>());
     });
 
   });