+++ mobx/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.0.7+4
 fixes:
 - shortened `1.asObservable()` to `1.obs()` (same for boolean, double, String) - [@subzero911](https://github.com/subzero911)
-- ObservableInt shortened to ObsInt (same for boolean, double, String)
+- removed experimental typedefs from `2.0.7`
 
 ## 2.0.7+3